fix test bug
