+++ src/Sekiban.Core/Documents/Pools/SortableIdConditionNone.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Documents.ValueObjects;
+namespace Sekiban.Core.Documents.Pools;
+
+public record SortableIdConditionNone : ISortableIdCondition
+{
+    public bool OutsideOfRange(SortableUniqueIdValue toCompare) => false;
+}
