+++ src/Sekiban.Core/Exceptions/SekibanEventRetrievalException.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Exceptions;
+
+public class SekibanEventRetrievalException(string message) : Exception(message), ISekibanException;
