+++ src/Sekiban.Core/Query/SingleProjections/Projections/IMemoryCacheContainer.cs
@@ -1,0 +1,7 @@
+using Sekiban.Core.Documents.ValueObjects;
+namespace Sekiban.Core.Query.SingleProjections.Projections;
+
+public interface IMemoryCacheContainer
+{
+    public SortableUniqueIdValue? SafeSortableUniqueId { get; }
+}
