+++ tests/Sekiban.Test.CosmosDb/Stories/CosmosEventOrderSpec.cs
@@ -4,9 +4,9 @@ using Sekiban.Test.Abstructs.Abstructs;
 using Xunit.Abstractions;
 namespace Sekiban.Test.CosmosDb.Stories;
 
-public class CosmosCustomerDbStoryBasic(
+public class CosmosEventOrderSpec(
     TestBase<FeatureCheckDependency>.SekibanTestFixture sekibanTestFixture,
-    ITestOutputHelper testOutputHelper) : CustomerDbStoryBasic(
+    ITestOutputHelper testOutputHelper) : EventOrderSpec(
     sekibanTestFixture,
     testOutputHelper,
     new CosmosSekibanServiceProviderGenerator());
