+++ tests/Sekiban.Test.Dynamo/Tests/EventOrderSpecDynamo.cs
@@ -4,9 +4,9 @@ using Sekiban.Test.Abstructs.Abstructs;
 using Xunit.Abstractions;
 namespace Sekiban.Test.Dynamo.Tests;
 
-public class CustomerDbStoryBasicDynamo(
+public class EventOrderSpecDynamo(
     TestBase<FeatureCheckDependency>.SekibanTestFixture sekibanTestFixture,
-    ITestOutputHelper testOutputHelper) : CustomerDbStoryBasic(
+    ITestOutputHelper testOutputHelper) : EventOrderSpec(
     sekibanTestFixture,
     testOutputHelper,
     new DynamoSekibanServiceProviderGenerator());
